backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
python-jose[cryptography]>=3.3.0
supabase>=2.3.4
psycopg2-binary>=2.9.10
slowapi>=0.1.9
pydantic>=2.6.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
httpx>=0.27.0
python-multipart>=0.0.9
playwright>=1.42.0
